+++ Branch-SDK/Branch-SDK/BNCPreferenceHelper.m
@@ -565,6 +565,32 @@ static NSString * const BRANCH_PREFS_KEY_ANALYTICS_MANIFEST = @"bnc_branch_analy
     }
 }
 
+- (NSArray<NSString*>*) URLBlackList {
+    @synchronized(self) {
+        id a = [self readObjectFromDefaults:@"URLBlackList"];
+        if ([a isKindOfClass:NSArray.class]) return a;
+        return nil;
+    }
+}
+
+- (void) setURLBlackList:(NSArray<NSString *> *)URLBlackList {
+    @synchronized(self) {
+        [self writeObjectToDefaults:@"URLBlackList" value:URLBlackList];
+    }
+}
+
+- (NSInteger) URLBlackListVersion {
+    @synchronized(self) {
+        return [self readIntegerFromDefaults:@"URLBlackListVersion"];
+    }
+}
+
+- (void) setURLBlackListVersion:(NSInteger)URLBlackListVersion {
+    @synchronized(self) {
+        [self writeIntegerToDefaults:@"URLBlackListVersion" value:URLBlackListVersion];
+    }
+}
+
 #pragma mark - Credit Storage
 
 - (NSMutableDictionary *)creditsDictionary {
@@ -775,7 +801,7 @@ static NSString * const BRANCH_PREFS_KEY_ANALYTICS_MANIFEST = @"bnc_branch_analy
 - (NSInteger)readIntegerFromDefaults:(NSString *)key {
     @synchronized(self) {
         NSNumber *number = self.persistenceDict[key];
-        if (number) {
+        if (number != nil) {
             return [number integerValue];
         }
         return NSNotFound;
@@ -848,7 +874,7 @@ NSURL* _Null_unspecified BNCCreateDirectoryForBranchURLWithSearchPath_Unthreaded
         if (success) {
             return branchURL;
         } else  {
-            BNCLogError(@"CreateBranchURL failed: %@ URL: %@.", error, branchURL);
+            NSLog(@"[branch.io] Info: CreateBranchURL failed: %@ URL: %@.", error, branchURL);
         }
     }
     return nil;
@@ -857,6 +883,7 @@ NSURL* _Null_unspecified BNCCreateDirectoryForBranchURLWithSearchPath_Unthreaded
 NSURL* _Nonnull BNCURLForBranchDirectory_Unthreaded() {
     NSArray *kSearchDirectories = @[
         @(NSApplicationSupportDirectory),
+        @(NSLibraryDirectory),
         @(NSCachesDirectory),
         @(NSDocumentDirectory),
     ];
@@ -879,7 +906,7 @@ NSURL* _Nonnull BNCURLForBranchDirectory_Unthreaded() {
             attributes:nil
             error:&error];
     if (!success) {
-        BNCLogError(@"Worst case CreateBranchURL error: %@ URL: %@.", error, branchURL);
+        NSLog(@"[io.branch] Error: Worst case CreateBranchURL error was: %@ URL: %@.", error, branchURL);
     }
     return branchURL;
 }