+++ Branch-TestBed/Branch-SDK-Tests/BNCApplication.Test.m
@@ -25,20 +25,10 @@
 
     BNCApplication *application = [BNCApplication currentApplication];
     XCTAssertEqualObjects(application.bundleID,                     @"io.branch.sdk.Branch-TestBed");
-    XCTAssertEqualObjects(application.teamID,                       @"R63EM248DP");
-    XCTAssertEqualObjects(application.applicationID,                @"R63EM248DP.io.branch.sdk.Branch-TestBed");
     XCTAssertEqualObjects(application.displayName,                  @"Branch-TestBed");
     XCTAssertEqualObjects(application.shortDisplayName,             @"Branch-TestBed");
     XCTAssertEqualObjects(application.displayVersionString,         @"1.1");
     XCTAssertEqualObjects(application.versionString,                @"1");
-    XCTAssertEqualObjects(application.pushNotificationEnvironment,  @"development");
-    XCTAssertEqual(application.keychainAccessGroups,                nil);
-    NSArray *domains = @[
-        @"applinks:bnc.lt",
-        @"applinks:bnctestbed.app.link",
-        @"applinks:bnctestbed.test-app.link"
-    ];
-    XCTAssertEqualObjects(application.associatedDomains, domains);
 }
 
 - (void) testAppDates {
@@ -73,22 +63,6 @@
             key:kBranchKeychainFirstInstalldKey
             error:nil];
     XCTAssertEqualObjects(application.firstInstallDate, firstInstallDate);
-}
-
-- (void) testIdentities {
-    if ([UIApplication sharedApplication] == nil) {
-        NSLog(@"No host application for BNCApplication testing!");
-        return;
-    }
-
-    BNCApplication *application = [BNCApplication currentApplication];
-    NSMutableDictionary * d =
-        [NSMutableDictionary dictionaryWithDictionary:application.deviceKeyIdentityValueDictionary];
-    XCTAssertTrue(d != nil);
-    [application addDeviceID:@"a" identityID:@"1"];
-
-    d[@"a"] = @"1";
-    XCTAssertEqualObjects(d, application.deviceKeyIdentityValueDictionary);
 }
 
 @end