+++ Branch.framework/Versions/A/Headers/BNCPreferenceHelper.h
@@ -49,6 +49,10 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory(void);
 @property (strong, atomic) NSString *referringURL;
 @property (strong, atomic) NSString *branchAPIURL;
 @property (assign, atomic) BOOL      limitFacebookTracking;
+@property (strong, atomic) NSDate   *previousAppBuildDate;
+
+@property (strong, atomic) NSArray<NSString*> *URLBlackList;
+@property (assign, atomic) NSInteger URLBlackListVersion;
 
 + (BNCPreferenceHelper *)preferenceHelper;